Update rec0vering2.sh
added and changed some echo's and added if statement to check if bootstrap was moved succesfully

+++ rec0vering2.sh
@@ -9,19 +9,27 @@ curl -L -o unc0verstrap.tar.lzma https://github.com/pwn20wndstuff/Undecimus/raw/
 
 if [[ -e /var/mobile/Documents/unc0verstrap.tar.lzma ]]
 then
+echo "Bootstrap downloaded... Moving Bootstrap (2/X)"
 mkdir /var/mobile/Documents/unc0verstrap
 mv /var/mobile/Documents/uc0verstrap.tar.lzma /var/mobile/Documents/unc0verstrap
 
 echo "Waiting for Bootstrap to be moved..."
+# Added sleep function to give time for the bootstrap to be moved.
 sleep 5
-echo "Ectracting Bootstrap..."
+if [[ -e /var/mobile/Documents/unc0verstrap/unc0verstrap.tar.lzma ]]
+then
+echo "Extracting Bootstrap... (3/X)"
 cd /var/mobile/Documents/unc0verstrap
-
 tar -xvf unc0verstrap.tar.lzma
 rm unc0verstrap.tar.lzma
+else
+echo "Error! : Bootstrap couldn't be moved."
+exit
+fi
 else
 echo "Error! : Couldn't download or find the bootstrap."
 exit
 fi
-echo "Scanning Filesystem for JB files (2/X)"
+echo "Scanning Filesystem for JB files (4/X)"
+
 fi
